requirements.txt
Flask==2.2.5
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-MongoEngine==1.0.0
Werkzeug==2.2.3
python-dotenv==1.1.0
